Feature: make mixing restarts if `mixing_restart` is smaller than drho (#3681)
* add a new member mixing_restart in charge_mixing

* modify eachiterinit() for LCAO

* modify hamilt2density() and eachiterfinish() for LCAO

* modify eachiterinit() for PW

* modify ESolver_KS::Run()

* change input.mixing_restart and GlobalV::MIXING_RESTART from int to double

* modify corresponding docs

* make some corrections

* delete useless block in ESolver_KS::Run()

* fix the bug to determine conv_elec

+++ source/module_base/global_variable.h
@@ -277,7 +277,7 @@ extern std::string of_kernel_file; // The name of WT kernel file.
 extern std::string MIXING_MODE;
 extern double MIXING_BETA;
 extern int MIXING_NDIM;
-extern int MIXING_RESTART;
+extern double MIXING_RESTART;
 extern double MIXING_GG0;
 extern bool MIXING_TAU;
 extern double MIXING_BETA_MAG;
