+++ source/module_io/parameter_pool.cpp
@@ -820,7 +820,7 @@ void input_parameters_set(std::map<std::string, InputParameter> input_parameters
     }
     else if (input_parameters.count("mixing_restart") != 0)
     {
-        INPUT.mixing_restart = *static_cast<int*>(input_parameters["mixing_restart"].get());
+        INPUT.mixing_restart = *static_cast<double*>(input_parameters["mixing_restart"].get());
     }
     else if (input_parameters.count("mixing_gg0") != 0)
     {
